backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1

torch==2.2.2
torchvision==0.17.2

pillow==10.3.0
opencv-python==4.13.0.92

numpy==1.26.4
gunicorn==21.2.0